Updates changing the Regex expression validation
Found dificulties using simple .matches method with regEx expressions. Changed to using the Pattern/Matcher classes in order to use RegEX on java.

Finishing up validation of first name as per rules:
Only letters, added a rule to show an error whenever an special character is found since special characters are not accepted on any of the fields.